Major refactor -> removed all nested index.tsx for flatter component structure.
Did a bit of digging around on open source repos, realized I was overnesting components.
