+++ src/components/services-carousel.tsx
