--- src/components/stack/StackGroupHorizontal/index.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-type Props = {
-  children: React.ReactNode;
-};
-
-const StackGroupHorizontal = ({ children }: Props) => {
-  return <div className="">{children}</div>;
-};
-
-export default StackGroupHorizontal;
