+++ src/hooks/use-low-power-mode.ts
